SpaceNet: convert to benchmark suite
