+++ torchgeo/datasets/spacenet/__init__.py
@@ -1,0 +1,26 @@
+# Copyright (c) Microsoft Corporation. All rights reserved.
+# Licensed under the MIT License.
+
+"""SpaceNet datasets."""
+
+from .base import SpaceNet
+from .spacenet1 import SpaceNet1
+from .spacenet2 import SpaceNet2
+from .spacenet3 import SpaceNet3
+from .spacenet4 import SpaceNet4
+from .spacenet5 import SpaceNet5
+from .spacenet6 import SpaceNet6
+from .spacenet7 import SpaceNet7
+from .spacenet8 import SpaceNet8
+
+__all__ = (
+    'SpaceNet',
+    'SpaceNet1',
+    'SpaceNet2',
+    'SpaceNet3',
+    'SpaceNet4',
+    'SpaceNet5',
+    'SpaceNet6',
+    'SpaceNet7',
+    'SpaceNet8',
+)
